added argument "no_hello_video" to prevent the build and install of hello_video

+++ install.sh
@@ -9,20 +9,27 @@ if [ "$(id -u)" != "0" ]; then
   exit 1
 fi
 
+
+
 echo "Installing dependencies..."
 echo "=========================="
 apt-get update
-apt-get -y install build-essential python-dev python-pip python-pygame supervisor git omxplayer
+apt-get -y install python-dev python-pip python-pygame supervisor git omxplayer
 
-echo "Installing hello_video..."
-echo "========================="
-git clone https://github.com/adafruit/pi_hello_video.git
-cd pi_hello_video
-./rebuild.sh
-cd hello_video
-make install
-cd ../..
-rm -rf pi_hello_video
+#check if hello_video should be installed
+if [ "$*" != "no_hello_video" ]
+then
+	echo "Installing hello_video..."
+	echo "========================="
+	apt-get -y install build-essential
+	git clone https://github.com/adafruit/pi_hello_video.git
+	cd pi_hello_video
+	./rebuild.sh
+	cd hello_video
+	make install
+	cd ../..
+	rm -rf pi_hello_video
+fi
 
 echo "Installing video_looper program..."
 echo "=================================="
